child process workers now attempt to terminate themselves via message
This causes workers run via `child_process.fork()` to respond to a special message by exiting.

On Windows especially, a subprocess killed via `SIGINT` cannot "clean up" after itself, and must exit immediately.  This has many implications, but importantly, it can cause `nyc` to fail to report code coverage on worker scripts.  A workaround is to listen for a "terminate" message from the parent, perform cleanup (if necessary), and call `process.exit()` itself.  `nyc` will then be successful in determining the coverage.

Since a child process worker may be very busy (e.g., in an infinite loop) and unable to _handle_ a message from its parent, `workerpool` now has a hardcoded timeout of `1000` ms; if the child does not shut down cleanly within this window, `workerpool` will ask the OS to signal a shutdown via `process.kill()` (which was the old behavior).

Short of adding `nyc` and asserting it calculates coverage, the current test suite does a reasonable job of avoiding regressions here.